pyproject.toml
[project]
name = "zepto-on-drugs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.14",
    "firebase-admin>=6.9.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "supabase>=2.16.0",
    "urllib3>=2.5.0",
    "uvicorn[standard]>=0.34.3",
]
